feat: pdf file preview also fixes s3 missing content-length header
